Update go-plugin, add support for reattach.
This updates the go-plugin dependency to 1.2.2, which brings in support
for a lot of things, but most importantly for this PR's purposes, it
brings in the ability for a server to set its own listener. This isn't
used on the Terraform side, but I figured it would be best for Terraform
and the SDK to be on the same version of go-plugin, and we need that
behavior on the SDK side, so I figured I'd update while I was here.

This also adds a TF_PROVIDER_SOFT_STOP environment variable which, when
set, will make Terraform shut down the GRPC server that the provider has
running instead of killing the process outright.

This also adds a TF_PROVIDER_REATTACH environment variable, which, when
set, will read go-plugin reattach information out of the environment
variable instead of starting a new process for the provider.

Both of these changes are in the pursuit of making Terraform providers
debuggable during normal operation (reattach) by allowing delve to start
the provider process and Terraform to connect to it, and making test
runs debuggable (soft stop instead of killing the process) by allowing a
single long running test process to have multiple Terraform commands run
against it by starting a server for each command, without the process
getting killed.